# projects/ecommerce_store/backend/backend_app/api/routes/cart_item.py
# GENERATED by the Backend Agent from the project graphs. Do not edit by hand: regenerate instead.
"""HTTP routes for service.cart_item. Route handlers only delegate to the application service."""
from __future__ import annotations

from typing import Literal
from uuid import UUID

from fastapi import APIRouter, Depends, Query, Response

from ...api_common import error_responses
from ...auth import authorize
from ...engine import ListQuery, Principal
from ..deps import get_services
from ...application.services import Services
from ..schemas.cart_item import CartItemCreateRequest, CartItemResponse, CartItemUpdateRequest


router = APIRouter(tags=["cart_item"])


@router.get(
    "/cart-items",
    operation_id="api.cart_item.list",
    summary="List Cart Items",
    description="List Cart Items.\n\nOperation `operation.cart_item.list`. Access: restricted (customer).",
    response_model=list[CartItemResponse],
    status_code=200,
    responses=error_responses(authenticated=True, restricted=True, has_path_id=False, mutating=False, rate_limited=False),
)
def api_cart_item_list(
    response: Response,
    limit: int = Query(100, ge=1, le=1000),
    offset: int = Query(0, ge=0),
    sort: str | None = Query(None, max_length=64, description="Attribute to sort by"),
    order: Literal['asc', 'desc'] = Query('asc'),
    principal: Principal | None = Depends(authorize("operation.cart_item.list")),
    services: Services = Depends(get_services),
):
    page = services.cart_item.list(principal, ListQuery(filters={}, search=None, sort=sort, order=order, limit=limit, offset=offset))
    response.headers['X-Total-Count'] = str(page.total)
    return page.items


@router.post(
    "/cart-items",
    operation_id="api.cart_item.create",
    summary="Create Cart Item",
    description="Create Cart Item.\n\nOperation `operation.cart_item.create`. Access: restricted (customer).",
    response_model=CartItemResponse,
    status_code=201,
    responses=error_responses(authenticated=True, restricted=True, has_path_id=False, mutating=True, rate_limited=False),
)
def api_cart_item_create(
    body: CartItemCreateRequest,
    principal: Principal | None = Depends(authorize("operation.cart_item.create")),
    services: Services = Depends(get_services),
):
    return services.cart_item.create(principal, body)


@router.delete(
    "/cart-items/{id}",
    operation_id="api.cart_item.delete",
    summary="Delete Cart Item",
    description="Delete Cart Item.\n\nOperation `operation.cart_item.delete`. Access: restricted (customer).",
    response_model=None,
    status_code=204,
    responses=error_responses(authenticated=True, restricted=True, has_path_id=True, mutating=True, rate_limited=False),
    response_class=Response,
)
def api_cart_item_delete(
    id: UUID,
    principal: Principal | None = Depends(authorize("operation.cart_item.delete")),
    services: Services = Depends(get_services),
):
    services.cart_item.delete(principal, id)
    return Response(status_code=204)


@router.patch(
    "/cart-items/{id}",
    operation_id="api.cart_item.update",
    summary="Update Cart Item",
    description="Update Cart Item.\n\nOperation `operation.cart_item.update`. Access: restricted (customer).",
    response_model=CartItemResponse,
    status_code=200,
    responses=error_responses(authenticated=True, restricted=True, has_path_id=True, mutating=True, rate_limited=False),
)
def api_cart_item_update(
    id: UUID,
    body: CartItemUpdateRequest,
    principal: Principal | None = Depends(authorize("operation.cart_item.update")),
    services: Services = Depends(get_services),
):
    return services.cart_item.update(principal, id, body)