# projects/ecommerce_store/backend/backend_app/api/schemas/cart_item.py
# GENERATED by the Backend Agent from the project graphs. Do not edit by hand: regenerate instead.
"""API request/response models for entity.cart_item (from api.json schemas)."""
from __future__ import annotations

from uuid import UUID

from pydantic import BaseModel, ConfigDict, Field


class CartItemResponse(BaseModel):
    """schema.cart_item: Cart Item"""

    model_config = ConfigDict(from_attributes=True)
    id: UUID
    user_id: UUID
    product_id: UUID
    quantity: int


class CartItemCreateRequest(BaseModel):
    """schema.cart_item.create: Create Cart Item request"""

    model_config = ConfigDict(extra="forbid", str_strip_whitespace=True)
    product_id: UUID
    quantity: int = Field(..., ge=-2147483648, le=2147483647)
    user_id: UUID | None = Field(None, json_schema_extra={'readOnly': True, 'x-server-owned': True})


class CartItemUpdateRequest(BaseModel):
    """schema.cart_item.update: Update Cart Item request"""

    model_config = ConfigDict(extra="forbid", str_strip_whitespace=True)
    quantity: int = Field(None, ge=-2147483648, le=2147483647)